Refactor FormatNumberFunctionIterator to use StaticContext for decimal formats and remove DecimalFormatRuntimeConfig Add condition to drop decimal-format information for all functions from static context that don't use it
